requirements.txt
git+git://github.com/Xaroth/libottdadmin2.git@master
marshmallow
flask-marshmallow
flask-sqlalchemy
sqlalchemy
marshmallow_enum
flask-smorest
Flask-Migrate
marshmallow-sqlalchemy
psycopg2
flask-cors
python-dotenv
flask-threads